xxx removed vmcontext
