+++ internal/pkg/testhelpers/consensus.go
@@ -2,14 +2,11 @@ package testhelpers
 
 import (
 	"context"
-	"math/big"
 	"testing"
 
 	bls "github.com/filecoin-project/filecoin-ffi"
 	"github.com/filecoin-project/go-filecoin/internal/pkg/block"
-	"github.com/filecoin-project/go-filecoin/internal/pkg/vm/abi"
 	"github.com/filecoin-project/go-filecoin/internal/pkg/vm/actor/builtin"
-	"github.com/filecoin-project/go-filecoin/internal/pkg/vm/actor/builtin/initactor"
 	cid "github.com/ipfs/go-cid"
 	"github.com/stretchr/testify/require"
 
@@ -144,20 +141,24 @@ func (ms testSigner) SignBytes(data []byte, addr address.Address) (types.Signatu
 
 // RequireActorIDAddress looks up an actor address in the init actor and returns the associated id address
 func RequireActorIDAddress(ctx context.Context, t *testing.T, st state.Tree, store vm.StorageMap, addr address.Address) address.Address {
-	processor := consensus.NewConfiguredProcessor(builtin.DefaultActors)
-	params, err := abi.ToEncodedValues(addr)
-	require.NoError(t, err)
+	// Dragons: why is this needed? delete
 
-	result, _, err := processor.CallQueryMethod(ctx, st, store, address.InitAddress, initactor.GetActorIDForAddressMethodID, params, address.Undef, nil)
-	require.NoError(t, err)
+	// processor := consensus.NewConfiguredProcessor(builtin.DefaultActors)
+	// params, err := abi.ToEncodedValues(addr)
+	// require.NoError(t, err)
 
-	idVal, err := abi.Deserialize(result[0], abi.Integer)
-	require.NoError(t, err)
+	// result, _, err := processor.CallQueryMethod(ctx, st, store, address.InitAddress, initactor.GetActorIDForAddressMethodID, params, address.Undef, nil)
+	// require.NoError(t, err)
 
-	idAddr, err := address.NewIDAddress(idVal.Val.(*big.Int).Uint64())
-	require.NoError(t, err)
+	// idVal, err := abi.Deserialize(result[0], abi.Integer)
+	// require.NoError(t, err)
 
-	return idAddr
+	// idAddr, err := address.NewIDAddress(idVal.Val.(*big.Int).Uint64())
+	// require.NoError(t, err)
+
+	// return idAddr
+
+	return address.Undef
 }
 
 // ApplyTestMessage sends a message directly to the vm, bypassing message
